requirements.txt
flask==2.3.2
pytest>=7.0.0
flake8>=6.0.0
black==23.3.0
pylint==2.17.4
autopep8==1.5.7
autoflake>=2.1.1
blinker==1.6.2
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
Werkzeug==3.0.1
Flask-Cors==4.0.1
Flask-Session==0.8.0
Bandit==1.7.0
python-dotenv==0.19.1
gunicorn==20.1.0
psycopg2-binary==2.9.1
python-dotenv==0.19.1